requirements.txt
python-dotenv==0.19.2
rich==11.0.0
youtube_dl==2021.6.6
